+++ SDL_VoxelEngine/SDL_VoxelEngine/Camera.h
@@ -1,12 +1,12 @@
 class Camera {
 public:
-	double x;
-	double y;
-	double z;
-	double rotX;
-	double rotY;
-	double rotZ;
-	double fovH;
-	double fovV;
-	Camera(double, double, double, double, double, double, double, double);
+	float x;
+	float y;
+	float z;
+	float rotX;
+	float rotY;
+	float rotZ;
+	float fovH;
+	float fovV;
+	Camera(float, float, float, float, float, float, float, float);
 };
